+++ server/api/apps/quiz_create/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class QuizCreateConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'api.apps.quiz_create'
