+++ server/api/apps/quiz_create/defines.py
@@ -1,0 +1,5 @@
+QUIZ_NAME_MAXLEN = 100
+QUIZ_TAG_MAXLEN = 15
+QUIZ_OBJECTIVE_MAXLEN = 15
+QUIZ_QUESTION_MAXLEN = 100
+QUIZ_ANSWER_MAXLEN = 100
