+++ server/api/apps/quiz_create/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from . import views
+
+# will change this later
+urlpatterns = [
+    path('', views.index, name='index'),
+]
