+++ server/api/apps/quiz_create/views.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+from django.http import HttpResponse
+
+
+# will change this later
+def index(request):
+    return HttpResponse("Hello, world.")
