--- server/quiz_create/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class QuizCreateConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'quiz_create'
